Convert input to C++ and add smooth movement
The movement system previously was kind of weird. It was for teleporting, but it looked like it was set up for smooth movement.
I changed it so that it still teleports the same way it did, but now if you hold the grip button and press the trigger at the same time it will move you in the direction of the laser at a speed proportional to how much you're pressing the trigger.
I also updated the instructions so that they explain the movement controls (as they are somewhat more complicated now)